+++ CHANGELOG.md
@@ -6,6 +6,7 @@ LUYA CHANGELOG
 
 **ISSUES**
 
+- `#696` Fixed bug where set relation method in ngrest model for checkbox realtions can handle not angular conform post data.
 - `#693` Added mailer component method adresses() function to assigne multiple recipients at the same time.
 - `#691` Create the module migration folder if the folder does not exists.
 - `#687` Fixed issue in StorageImporter and empty storage folder.